ibm-data-engineering-course/pyproject.toml
[tool.poetry]
name = "ibm-data-engineering-course"
version = "0.1.0"
description = ""
authors = ["luikymagno <luikymagno@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^2.0.2"
postgres = "^4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
